Fix MaskingFunctionTester max timeInMillisToDay to conform with changes made in STAR-692

+++ test/unit/org/apache/cassandra/cql3/functions/masking/MaskingFunctionTester.java
@@ -110,7 +110,7 @@ public abstract class MaskingFunctionTester extends CQLTester
                 case DATE:
                     testMaskingOnAllColumns(type,
                                             SimpleDateSerializer.timeInMillisToDay(2),
-                                            SimpleDateSerializer.timeInMillisToDay(Long.MAX_VALUE));
+                                            SimpleDateSerializer.timeInMillisToDay(Integer.MAX_VALUE));
                     break;
                 case DURATION:
                     testMaskingOnNotKeyColumns(type, Duration.newInstance(1, 2, 3), Duration.newInstance(3, 2, 1));
